#5726 Updated code structure
